// lib/types/enum.ts
export enum MethodEnum {
  DFS = "DFS",
  BFS = "BFS",
  GBFS = "Greedy Best-First Search",
  DFSH = "DFS Heuristic",
  ASTAR = "A*",
  JPS = "Jump-Point Search",
}

export enum HeuristicEnum {
  Manhattan = "Manhattan",
  Euclidean = "Euclidean",
  Chebyshev = "Chebysev",
}

export enum YieldEnum {
  SEARCH = "search",
  SCAN = "scan",
  PATH = "path",
  ERROR = "error",
}